requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.7
pandas==2.2.0
openpyxl==3.1.2
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0